Crate gst

Re-exports

pub use self::appsink::AppSink;

pub use self::bus::Bus;

pub use self::mainloop::MainLoop;

Modules

appsink
Easy way for applications to extract samples from a pipeline.

bus

ffi

mainloop

Structs

AppSrc

Bin
GstBin is an element that can contain other GstElement, allowing them to be managed as a group. Pads from the child elements can be ghosted to the bin, see GstGhostPad. This makes the bin look like any other elements and enables creation of higher-level abstraction elements.

Buffer

BufferPool

Caps

Element

Error

Iter

MiniObject

Object

Pad

Pipeline
A GstPipeline is a special GstBin used as the toplevel container for the filter graph. The GstPipeline will manage the selection and distribution of a global GstClock as well as provide a GstBus to the application.

PlayBin

Ref
Represents a reference to a gstreamer object equivalent to gst_object_ref and gst_mini_object_ref it auto unrefs when going out of scope

Sample

Structure

VideoComponent

VideoFrame

VideoPlane

Enums

Map

Message

Traits

FromGValue

Transfer

Functions

filename_to_uri

init

uri_get_protocol

Type Definitions

MapInfo

Result

VideoInfo