Struct gst::AppSrc

pub struct AppSrc { /* fields omitted */ }

Methods

impl AppSrc

pub fn new(name: &str) -> Option<AppSrc>

pub fn new_from_element(element: Element) -> AppSrc

pub fn set_caps(&mut self, caps: &Caps)

Set the capabilities on the AppSrc. After calling this method, the source will only produce caps that match caps. Once caps is set, the caps on the buffers MUST either match the caps OR be left unspecified.

Before operating an AppSrc, the caps property MUST be set to fixed caps describing the format of the data that will be pushed with appsrc EXCEPT when pushing buffers with unknown caps, in which case no caps should be set. This is typically true of file-like sources that push raw byte buffers.

pub fn caps(&self) -> Option<Caps>

pub fn latency(&self) -> (u64, u64)

pub fn push_buffer(&mut self, buffer: Buffer) -> GstFlowReturn

pub fn end_of_stream(&mut self) -> GstFlowReturn

pub unsafe fn gst_appsrc(&self) -> *const GstAppSrc

pub unsafe fn gst_appsrc_mut(&mut self) -> *mut GstAppSrc