Trait gst::Transfer

pub trait Transfer<PtrType = GstElement> {
    unsafe fn transfer(self) -> *mut PtrType;
}

Required methods

unsafe fn transfer(self) -> *mut PtrType

Consumes the current object and transfers ownership of the raw pointer Used to transfer ownership to ffi functions, should be used when an ffi function expects full transfer of an object to avoid the original object to be unreferenced in the process

Loading content...